вход и регистрация пределаны, добавлен класс MyAuthorization, удалены лишние классы
